Change PartitionKey from u32 to u64
This commit changes the Partitionkey to be a u64 instead of
a u32.

Important: This change will break all existing Restate clusters
since the existing state will no longer be compatible with new
PartitionKey type.

This fixes #714.

+++ src/storage_rocksdb/tests/status_table_test/mod.rs
@@ -73,7 +73,7 @@ async fn populate_data<T: StatusTable>(txn: &mut T) {
     .await;
 
     txn.put_invocation_status(
-        &ServiceId::with_partition_key(u32::MAX, "svc-u64", "key-0"),
+        &ServiceId::with_partition_key(u64::MAX, "svc-u64", "key-0"),
         invoked_status(uuid_str("218756fa-3f7f-7854-a76b-42c59a3d7f2d")),
     )
     .await;
@@ -150,9 +150,9 @@ async fn verify_lookup_by_invocation_id_not_found<T: StatusTable>(txn: &mut T) {
 }
 
 async fn verify_last_partition_all_svc_with_status_invoked<T: StatusTable>(txn: &mut T) {
-    let stream = txn.invoked_invocations(4000..=u32::MAX);
+    let stream = txn.invoked_invocations(4000..=u64::MAX);
     let expected = vec![ServiceInvocationId::with_service_id(
-        ServiceId::with_partition_key(u32::MAX, "svc-u64", "key-0"),
+        ServiceId::with_partition_key(u64::MAX, "svc-u64", "key-0"),
         uuid_str("218756fa-3f7f-7854-a76b-42c59a3d7f2d"),
     )];
